use sqlite to store node lookup table
While parsing the OSM XML file, we first need to index all the nodes, so that when parsing the ways we can assign geographic coordinates to all points that comprise a way. As there might be millions of nodes, they can often not be all stored in memory - so we use sqlite database with an index built on the id column for fast node lookup. Like this, it should be possible to parse very big OSM data files with acceptable memory usage.

 During my testing on the Czech Republic OSM data file, I got from 9.7 GB with the old in-memory node dictionary down to about 3.5 GB (which is strange in itself - it should still be much less).